requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.30.0
black==25.11.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
fastapi==0.121.2
flake8==7.3.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
isort==7.0.0
librt==0.6.3
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy==1.19.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
psycopg2-binary==2.9.11
pwdlib==0.3.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.20
pytokens==0.3.0
PyYAML==6.0.3
redis==7.0.1
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
